Modification of initial application, taking into account different parameters at startup
This is a new version of the calculator.
It provides new parameters to extend the generated data. You can now add information about the sunrise and sunset,
If the energy community exists, you should retrieve the obis codes calculated by the infrastructure provider. These are the codes 1-65:1.29.2 for the consumers, as well as the 1-65:2.29.2 for the producers. These values are already calculated by the metering company to take into account only the energy used internally by the energy community.
Check out my youtube playlist (in luxemburgish)  in order to get more information : 
https://www.youtube.com/watch?v=VdAfy_WvskM&list=PL0jofPMXMFcJ5fyEiPee74Hj4TCjClsaQ